# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vertex AI Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads with optional filters
      parameters:
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [new, drafted, failed, skipped]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLeads"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a single lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /leads/bulk-import:
    post:
      operationId: bulkImportLeads
      tags: [leads]
      summary: Bulk import leads from parsed CSV/XLSX data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /leads/{id}/retry-draft:
    post:
      operationId: retryLeadDraft
      tags: [leads]
      summary: Retry draft creation for a failed lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftResult"

  # ─── TEMPLATES ───────────────────────────────────────────────────────────────
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: List email templates for current user
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /templates/{id}/duplicate:
    post:
      operationId: duplicateTemplate
      tags: [templates]
      summary: Duplicate a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"

  # ─── AI ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: ["integer", "null"]
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        vehicle:
          type: ["string", "null"]
        route:
          type: ["string", "null"]
        pickup:
          type: ["string", "null"]
        delivery:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, drafted, failed, skipped]
        gmailDraftId:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, email, status, createdAt, updatedAt]
    LeadInput:
      type: object
      properties:
        campaignId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        vehicle:
          type: string
        route:
          type: string
        pickup:
          type: string
        delivery:
          type: string
        price:
          type: string
        notes:
          type: string
      required: [name, email]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        vehicle:
          type: string
        route:
          type: string
        pickup:
          type: string
        delivery:
          type: string
        price:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [new, drafted, failed, skipped]
    PaginatedLeads:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    BulkLeadImportInput:
      type: object
      properties:
        campaignId:
          type: ["integer", "null"]
        leads:
          type: array
          items:
            $ref: "#/components/schemas/LeadInput"
      required: [leads]
    BulkImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        duplicates:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [imported, skipped, duplicates, errors]
    DraftResult:
      type: object
      properties:
        success:
          type: boolean
        gmailDraftId:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
      required: [success]

    # ─── TEMPLATES ───────────────────────────────────────────────────────────
    EmailTemplate:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, subject, body, isDefault, createdAt, updatedAt]
    EmailTemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        isDefault:
          type: boolean
      required: [name, subject, body]
    EmailTemplateUpdate:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        isDefault:
          type: boolean

    # ─── AI ──────────────────────────────────────────────────────────────────
...